requirements.txt
flask>=3.0,<4
gspread>=6.0
google-auth>=2.0
python-dotenv>=1.0
pytz>=2024.1
